HotFix 20092020_10:59
Awaiting for user to be gotten and added AllowedHosts security.

+++ chat/consumers.py
@@ -1,3 +1,4 @@
+from channels.auth import get_user
 from channels.db import database_sync_to_async
 
 from core.consumers import HandlerJsonWebsocketConsumer
@@ -17,8 +18,7 @@ class ChatConsumer(HandlerJsonWebsocketConsumer):
         await super().disconnect(code)
 
     @database_sync_to_async
-    def register_message(self, author, chat_id, message):
-        author_id = author.id
+    def register_message(self, author_id, chat_id, message):
         message = models.ChatMessage.objects.create(
             author_id=author_id,
             chat_id=chat_id,
@@ -32,26 +32,33 @@ class ChatConsumer(HandlerJsonWebsocketConsumer):
         await self.channel_layer.group_add(self.chat_group_name, self.channel_name)
 
     async def send_message(self, content):
-        user = self.scope['user']
-        chat_id = content['chat']
-        msg_text = content['message']
-        message = await self.register_message(user, chat_id, msg_text)
-        data = serializers.ChatMessageSerializer(message).data
-        func = 'group_send_message'
-        content = {
-            'type': f'{func}',
-            'message': data
-        }
+        user = await get_user(self.scope)
 
-        try:
-            await self.channel_layer.group_send(self.chat_group_name, content)
-        except TypeError:
-            # We send error message with serialized message and remove
-            await self.send_json(
-                {'error': 'We couldn\'t send your message', 'message': data},
+        if user.is_authenticated:
+            chat_id = content['chat']
+            msg_text = content['message']
+            message = await self.register_message(user.id, chat_id, msg_text)
+            data = serializers.ChatMessageSerializer(message).data
+            func = 'group_send_message'
+            content = {
+                'type': f'{func}',
+                'message': data
+            }
+
+            try:
+                await self.channel_layer.group_send(self.chat_group_name, content)
+            except TypeError:
+                # We send error message with serialized message and remove
+                await self.send_json(
+                    {'error': 'We couldn\'t send your message', 'message': data},
+                    close=True
+                )
+                message.delete()
+        else:
+            self.send_json(
+                {'error': 'User is not authenticated.'},
                 close=True
             )
-            message.delete()
 
     async def group_send_message(self, content):
         message = content['message']
